funds: show transfer error when breez unable to pay for "funds added"
In this commit we update the FundStatusReply to contain
TRANSFER_ERROR field that will be used for the error that is returned
from breez in case it can't pay in lightning for the funds added by the
user.

In addition we populate this field with error returned from breez when
trying to get the payment. If all confirmed addresses contains error,
We return FundStatusReply_TRANSFER_ERROR as status and populate
the error message in the ErrorMessage field.