sync: rate limit sync job for every 10 minutes.
In this commit we make sure the sync job won't run more than 1 time
every 10 minutes.
This is important because there could be various triggers that triggers
the job (notifications, local scheduling etc...) and the way these
triggers are aggregated depedns on the OS. This commits ensures
there won't be a case where multiple jobs will run at the same point
In time, causing problems at resources in mobile.