invoice: Use one function to add invoice.
This commit removes the differentiation in the code of creating
"Standart Invoice" and breez invoice.
Instead there is one function AddInvoice that consolidates both
use cases depends on the data given. If the payer/payee data is given
the it is encoded texst wise into the invoice, if not then only the
description is added.