sync: introduce BreachWatcher.
In this commit a new struct called BreachWatcher is pushed.
This is intended to be used as part of the sync job to periodically scan
using neutrino ChainService for any breach within the user's channels.
If a breach is to be detected then the user should be warn to open the
app.