sync: use earlier channel block as initial start
This commit handles the case where the job runs first time.
In that case the db doesn't contain the starting point for the scan.
Therefore we use the earlier channel block height instead.
Also a small optimization was added to skip the rescan in case there are
no opened channels.